+++ src/components/common/Tab/Tab.tsx
@@ -1,10 +1,10 @@
 import { TAB_CONSTANTS } from '@/constants/tab';
 import { useTabContext } from '@/hooks/useTabContext';
-import Theme from '@/styles/Theme';
 import { getKeyByValue } from '@/utils/getKeyByValue';
-import styled from '@emotion/styled';
 
 import { useEffect } from 'react';
+
+import { SeparatorStyled, TabContainerStyled, TabItemStyled } from './Tab.style';
 
 interface TabItemProps {
   title: string;
@@ -18,29 +18,6 @@ interface TabProps {
   defaultTab?: string;
   tabItems: TabItemProps[];
 }
-
-const TabContainerStyled = styled.div<{ maxWidth: string }>`
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-  max-width: ${(props) => `${props.maxWidth}rem`};
-  width: 100%;
-`;
-
-const TabItemStyled = styled.div<{ isActive: boolean; width: string }>`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: ${(props) => `${props.width}rem`};
-  height: 2rem;
-  color: ${(props) => (props.isActive ? `${Theme.color.activeColor}` : `${Theme.color.black}`)};
-  cursor: pointer;
-`;
-
-const SeparatorStyled = styled.div`
-  height: 1.5rem;
-  border-left: 0.07rem solid;
-`;
 
 const Tab = ({ active, maxWidth, defaultTab, tabItems }: TabProps) => {
   const { activeTab, setActiveTab } = useTabContext();
